requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.1
tensorflow==2.13.0
pytorch==2.0.1
statsmodels==0.14.0

# Data processing
pybaseball==2.2.5
sportsipy==0.6.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.8.5
asyncio==3.4.3
sqlalchemy==2.0.19
alembic==1.11.1
psycopg2-binary==2.9.6

# API and web
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.24.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development tools
black==23.7.0
isort==5.12.0
flake8==6.0.0
mypy==1.4.1
pre-commit==3.3.3

# Documentation
sphinx==7.1.1
sphinx-rtd-theme==1.2.2

# Deployment
gunicorn==21.2.0
docker==6.1.3 